Revert "Containers run as a USER should have UID in /etc/passwd"
This reverts commit ca4b09f718b08796d5f87778e73d2e8fad11162a.
